feat: add dark theme for example project
