+++ lib/src/editor/editor_component/service/ime/delta_input_impl.dart
@@ -3,3 +3,5 @@ export 'delta_input_on_delete_impl.dart';
 export 'delta_input_on_insert_impl.dart';
 export 'delta_input_on_non_text_update_impl.dart';
 export 'delta_input_on_replace_impl.dart';
+export 'non_delta_input_service.dart';
+export 'text_input_service.dart';
